requirements.txt
streamlit
pandas
python-dotenv
pathway
google-generativeai
